#1 Fix review issues: move db info output to separated class
